+++ controllers/doguSupportManager.go
@@ -8,7 +8,6 @@ import (
 	"github.com/cloudogu/k8s-host-change/pkg/alias"
 
 	k8sv1 "github.com/cloudogu/k8s-dogu-operator/api/v1"
-	"github.com/cloudogu/k8s-dogu-operator/controllers/limit"
 	"github.com/cloudogu/k8s-dogu-operator/controllers/resource"
 
 	appsv1 "k8s.io/api/apps/v1"
@@ -38,7 +37,7 @@ type doguSupportManager struct {
 
 // NewDoguSupportManager creates a new instance of doguSupportManager.
 func NewDoguSupportManager(client client.Client, cesRegistry registry.Registry, eventRecorder record.EventRecorder) *doguSupportManager {
-	resourceGenerator := resource.NewResourceGenerator(client.Scheme(), limit.NewDoguDeploymentLimitPatcher(cesRegistry), alias.NewHostAliasGenerator(cesRegistry.GlobalConfig()))
+	resourceGenerator := resource.NewResourceGenerator(client.Scheme(), resource.NewRequirementsGenerator(cesRegistry), alias.NewHostAliasGenerator(cesRegistry.GlobalConfig()))
 	return &doguSupportManager{
 		client:            client,
 		doguRegistry:      cesRegistry.DoguRegistry(),